fix(schema): update resource fields to use BigInt
- Changed `Int4` fields to `BigInt` in `resource_generation` schema for better scalability.
- Updated Rust structs, schema definitions, and SQL migration to reflect the change.
- Ensures proper support for larger values in resource calculations.

+++ migrations/2024-05-01-074658_resource_generation/up.sql
@@ -1,14 +1,14 @@
 CREATE VIEW resource_generation AS
 SELECT pb.player_id,
-       SUM(br.population)    as population,
-       SUM(br.food)          as food,
-       SUM(br.wood)          as wood,
-       SUM(br.stone)         as stone,
-       SUM(br.gold)          as gold,
-       SUM(br.food_acc_cap)  as food_acc_cap,
-       SUM(br.wood_acc_cap)  as wood_acc_cap,
-       SUM(br.stone_acc_cap) as stone_acc_cap,
-       SUM(br.gold_acc_cap)  as gold_acc_cap
+       SUM(br.population)::bigint    as population,
+       SUM(br.food)::bigint          as food,
+       SUM(br.wood)::bigint          as wood,
+       SUM(br.stone)::bigint         as stone,
+       SUM(br.gold)::bigint          as gold,
+       SUM(br.food_acc_cap)::bigint  as food_acc_cap,
+       SUM(br.wood_acc_cap)::bigint  as wood_acc_cap,
+       SUM(br.stone_acc_cap)::bigint as stone_acc_cap,
+       SUM(br.gold_acc_cap)::bigint  as gold_acc_cap
 FROM player_building pb
          LEFT JOIN public.building_resource br
                    ON pb.building_id = br.building_id
